+++ ChangeLog.md
@@ -4,6 +4,11 @@
 
 ## Upcoming Release
 
+Blob:
+
+- Fixed issue of download 0 size blob with range > 0 should report error. (issue #2410)
+- Fixed issue of download a blob range without header x-ms-range-get-content-md5, should not return content-md5. (issue #2409)
+
 ## 2024.06 Version 3.31.0
 
 General: